pyproject.toml
[tool.poetry]
name = "orchs-domain-discovery-2"
version = "0.1.0"
description = "Domain discovery was so good we released sequel."
authors = ["matej smycka <smycka@ics.muni.cz>"]
license = ""
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.7"
sqlalchemy = "^2.0.31"
polars = "^1.2.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
mypy = "^1.3.0"
isort = "^5.12.0"
pyproject-flake8 = "^6.0.0.post1"
pytest = "^7.4.0"
ruff = "^0.1.0"

[tool.ruff]
line-length = 120
select = [
  "E",   # pycodestyle
  "F",   # pyflakes
  "UP",  # pyupgrade,
  "I",   # isort
]

